requirements.txt
alembic
annotated-types==0.7.0
anyio==4.9.0
certifi==2025.4.26
colorama==0.4.6
h11==0.16.0
httpcore==1.0.9
httpx==0.28.1
idna==3.10
iniconfig==2.1.0
packaging==25.0
pathvalidate==3.3.1
pluggy==1.5.0
pydantic==2.11.4
pydantic_core==2.33.2
python-dotenv==1.1.0
python-telegram-bot==22.0
PyYAML==6.0.2
sniffio==1.3.1
typing-inspection==0.4.0
typing_extensions==4.13.2
yt-dlp==2026.2.4
